+++ Serwer/SerwerApplication/Team.cpp
@@ -26,8 +26,8 @@ void Team::resetPosition()
 	int i = 1;
 	for(Player player : players)
 	{
-		int px = player.getGame()->getScreenW() / 2 + this->pitchHalf * this->pGame->getScreenW() / 4;
-		int py = (player.getGame()->getScreenH() - player.getGame()->getPitchH()) / 2 + i * player.getGame()->getPitchH() / 4;
+		float px = player.getGame()->getScreenW() / 2 + this->pitchHalf * this->pGame->getScreenW() / 4;
+		float py = (player.getGame()->getScreenH() - player.getGame()->getPitchH()) / 2 + i * player.getGame()->getPitchH() / 4;
 		player.setMove(Vector2D(0, 0));
 		player.setPosition(Vector2D(px, py));
 		i++;
